separate the routes from index, and created db.js where mongodb connection, and also created routes folder for separated routes
