+++ middlewares/authMiddleware.js
@@ -1,0 +1,18 @@
+const jwt = require("jsonwebtoken");
+
+const verifyToken = (req, res, next) => {
+    if (!req.headers.authorization) {
+        return res.status(401).send({ message: 'unauthorize access' });
+    }
+
+    const token = req.headers.authorization.split(' ')[1]
+    jwt.verify(token, process.env.ACCESS_TOKEN, (err, decoded) => {
+        if (err) {
+            return res.status(401).send({ message: 'unauthorize access' });
+        }
+        req.decoded = decoded;
+        next();
+    })
+};
+
+module.exports = verifyToken;
